requirements.txt
absl-py==1.0.0
aniso8601==9.0.1
argcomplete==1.12.3
argon2-cffi @ file:///C:/ci/argon2-cffi_1613037959010/work
astunparse==1.6.3
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.entry-points-selectable==1.1.1
bleach @ file:///tmp/build/80754af9/bleach_1628110601003/work
cachetools==4.2.4
certifi==2021.10.8
cffi @ file:///C:/ci/cffi_1636542074879/work
charset-normalizer==2.0.7
click==8.0.3
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.11.0
debugpy @ file:///C:/ci/debugpy_1637073815078/work
decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
distlib==0.3.3
entrypoints==0.3
filelock==3.4.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
flask-restplus==0.13.0
flatbuffers==2.0
fonttools==4.32.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
imageio==2.12.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916848487/work
ipykernel @ file:///C:/ci/ipykernel_1633545574250/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1635944310712/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous==2.0.1
jedi @ file:///C:/ci/jedi_1611315808330/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
joblib==1.1.0
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1637148478538/work
jupyter-core @ file:///C:/ci/jupyter_core_1636537213213/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe @ file:///C:/ci/markupsafe_1621528314575/work
matplotlib==3.5.1
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune==0.8.4
mysqlclient==2.1.0
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1624479163777/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx==2.8
notebook @ file:///C:/ci/notebook_1637143794296/work
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1625611678980/work
pandas==1.4.2
pandocfilters @ file:///C:/ci/pandocfilters_1605102497129/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
pipenv==2021.11.23
pipx==0.16.4
platformdirs==2.4.0
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1637050397234/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent @ file:///C:/ci/pyrsistent_1636111468851/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyWavelets==1.3.0
pywin32==228
pywinpty==0.5.7
pyzmq @ file:///C:/ci/pyzmq_1628276144921/work
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
split-folders==0.5.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
threadpoolctl==3.1.0
tifffile==2022.4.22
tornado @ file:///C:/ci/tornado_1606942392901/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.0.0
urllib3==1.26.7
userpath==1.7.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.13.3
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
